All tests and notebooks works
All unitary tests have been made to work. There was an issue in which the DummyEngine were not registered properly when in the test_ file. As such, dummy_engine.py were created in the unitary test folders so that DummyEngine could properly be registered using a decorator
Changed the notebooks to make them works with the new code architecture. Rmk : before using block analysis in a notebook, a fake problem has to be ran using a yaml, otherwise modules are not registered.
All integration tests work