Partial change to the definition of y_ratio
y_ratio is now defined as vector to accomodate for multiple engines on each wing. The input vector has to have a ize of ten, each unused element of the vector must be put to a strictly negative value !
Change were made in the geometry, CG computation, HQ computation, loads computation and proper plotting of multiple engines.
OpenVSP remains but will be in a separate commit