Submodels registered for geometry
Group submodels have been registered for geometry. Important point is that everything that is linked to how we size the fuselage (Fixed fuselage length or Fixed tail distance) has not been registered as a submodels as it is linked to the module itself. In consequence the choice of how we size the fuselage is done with the right choice of MODULE not SUBMODEL