Separation of wing area model
In response to issue #133 the wing area models are now separated, and called inside a new group that always uses a geometric and an aeroydnamic constaints. More constraints can be envisionned in the future. Submodels are used for this and the simple computations for both constraints were registered as default submodels. Also, propeller performances computation were removed from some OAD process to save time in the future.