Shape by conn added for wing loads computation
The variables with shape_by_conn which caused issue in the previous commit are now corrected. It seems like if variables have copied_shape they must all be called in components of the same groupe (example of structural loads). This should be the end of the transformation of old variables with shape by conn (some are left but it was deemed reasonable to leave them as is).