Propeller effective efficiency
Computed the ratio by which the efficiency must be multiplied to get the effective efficiency due to the additional drag of cowling behind the propeller. Computed but not used just yet. Expecting some additionnal loops due to the fact that the propulsion wrapper is now going to use these ratio and that the wrapper will be called before these are computed. Runtime should be monitored. Can be linked to issue #80 for possible solution.