Propulsion change for geometry computation
Modules and computation that depend directly on the propulsion wrapper has been changed to be isolated and registered as submodels. Nacelle compitation have been separated based on that and useless call to the propulsion wrapper have also been discarded. OAD process using VLM have also been separated to correct a bug that caused NaN to appear in input file when run in the same pytest