Post processing functions for props and API
Added post-processing function to print the efficiency maps for the propeller. The way to use those is presented in the second notebook. Also created an API for post-processing so all imports can be centralized. This caused a few bugs that are now fixed and tested. Also rewrote some part of the notebook to better make use of that api as well as extending the use of FAST-OAD api