Derivative of sideforce with respect to side slip
Computation of the derivative of sideforce with respect to side slip by summing the contribution of wing, vt and fuselage. Since Cy_beta for the VTP is now computed on its own, it was added as an input in the update VT area component.
Also introduced dihedral on the wing for the computation of the contribution of the wing. For now it is only used there and should be included more thoroughly in subsequent versions of the code