Derivative of yaw moment wrt sideslip
Computation of the derivative of yaw moment with respect to an sideslip angle by assuming only the contribution the fuselage and VTP is relevant. This coefficient however depends on the aircraft CG, so same remark as in 1608146 holds. Further investigation on why the target is not precisely equal to the output should be led