Two separate AoA inputs
Those aerodynamic coefficient computation are destined to be ran in the same FAST-OAD module both in high speed and low-speed which made it necessary to have two separate AoA inputs for the module that required one, which is what is done here. It also leaves more freedom for the users.