Removed mission solver
Removed the solver that used to solve the cruise mission so that the range of the mission was equal to the sum of the ranges of each parts. This convergence is now handled by the global solver with no difference on the result. This however implies that if the mission is to be ran on its own, it should be done with the addition of a solver.