Weight of High wing aircraft
A empirical method for the estimation of the weight of the aircraft was added based on the Cessna method desribed in Roskam. This method however gives very small results for low wing aircraft so a warning was added in the submodel in question.
The semi analytical method was also changed to account for heavier wing-fuselage connection.
This was all tested with very approximate value of the P68.
The high wing will remain a work in progress maybe until the next release so not too much effort was put in the validation of the model.