Addition of twist generation and dihedral
Testing the change in dihedral
Also added a method in VLM which enable to take the twist into account, there is however a little issue: since the lift distribution is computed at a small angle of attack and is then scale up and down when needed at different Cl, having negative lift at some point of the span will create false results when scaling.
This can happen when twist is added as even small twist can have very big impact at low AoA.
This will be solved by changing the reference used for the lift distribution which is returned to the one at AOA. This has however the disadvantage of requiring us to output the cl_ref and reusing it everywhere else in the code.