Change in the reference lift curve
The reference lift curve is now the one computed at INPUT_AOA instead of the one computed at 0.0 deg. It makes more sense as we primarily use the scaling at high angle of attack and it is going to spare use some trouble with the inclusion of twist as small twist at low AOA can significantly affect the distribution as mentioned in previous commits. Also changed the data file for the unitary tests of affected modules but output value have not been changed yet.