Addition of high wing in OpenVSP
Changed the value of z_wing that was given to the script to the input from the process instead of a value computed inside the file.
For VLM, since the downwash and dynamic pressure losses are computed empirically, the change in z position has no impact. Next come the verification of unit tests and intergration tests