Updated list_inputs_metadata
With the new version of FAST-OAD-core and openmdao the formerly used list_variables does not return the metadata anymore because the from_problem() function does not return them. To counter that, we are now using the  OpenMDAO method get_io_metadata and ensure to sort out duplicate. It should work now.