Numpy latest compatibility
OpenMDAO tends to transform float into 1 element array which was then fed into a list for an interpolation fucntion causin a shape mismatch. This is corrected by ensuring that the altitude in the init is an array and taking its items. This was the way used in the turboprop actually.