Propeller computation change
Wasn't part of the original PR but still made some change. We were unnecessarily extensive on how many propeller pitch we wre testing. Now we only do it once per degree before it was more every third of degree. Also attempted to provide smart initial values to the induced speed but it makes the low speed not converge :frowning: