Turboprop thermodynamic equations rewrite
Rewrote the formula necessary to compute the turboprop geometry, fuel consumption and maximum thrsut under the form of an OpenMDAO problem. The rationale behind it is that it seems to be quicker to solve it like this which would reduce the computation time of what is right the longest module in FAST-OAD_CS23. Now, what's left to do is implement the use of these components (whose validity were checked in another project) and optimize the implementation, using results caching or a smarter breakdown of the maps