Nested thrust max problems
The 4 different problems that will help computing the max thrust at all altitudes have now been implement as was done for the sizing of the turboprop. Note that a problem for the case where propeller thrust is limiting has been added. Their use in the max_thrust function has not yet been implemented !