Max thrust rework
Reworked the max thrust function so that it uses the nested problem. The main assumption is that, if limits are reached, they are reached in this order : power, opr, itt, propeller thrust.
Also added a small cache to the max thrust function for cases like cruise where altitude and speed don't vary