Nested fuel consumption problem
The 2 different problems that will help computing the fuel consumed have now been implement as was done for the sizing of the turboprop. There is a problem with a linesearch algorithm that can be used but will not be favored if possible. There implementation in the sfc function have not yet been implemented