Map construction rework
Reworked the cnstruction of the sfc map for the turboprop and checked tested, only minor difference on predicted sfc with new map generation for a considerable speedup (7m to 30s). Now we will test and rework the reading of those maps. Btw `scipy.interpolate.interp2d` is gonna be deprecated so it was changed