First attempt
It works very well during the unit test and the first few loops of the oad processs of the TBM900 but for some reason from the thrist iteration onward it iterated between two value and never converges. In the case of the TBM it is not a really a problem since the other criteria is the most constraining but it's not generic enough that it can be merged into the main. Still the gain in time makes it worth pursuing