Number of points as option
I realized that what was causing the erratic behaviour of the solver was the number of points and by decreasing it, faster convergence was obtained though at the cost of slightly different result. Consequently, that number was set as an option. However, I still can't make the line with "wing_area" disapear from the output file. I've narrowed down that it is linked, for some reasons, with the promotion of the DistanceToMFWForUpdate() component but can't change it